Refactor 'Transaction' related classes into 'transaction' package
Previously, package name was spelt incorrectly. (i.e. 'transcation' instead of 'transaction')
